fix: add markdown image preview overlay
