+++ frontend/src/components/ui/ImagePreviewOverlay.tsx
@@ -1,0 +1,85 @@
+"use client";
+
+import { useEffect } from "react";
+import { createPortal } from "react-dom";
+import { ExternalLink, X } from "lucide-react";
+
+interface ImagePreviewOverlayProps {
+  src: string;
+  title: string;
+  onClose: () => void;
+  onImageError: () => void;
+}
+
+export default function ImagePreviewOverlay({
+  src,
+  title,
+  onClose,
+  onImageError,
+}: ImagePreviewOverlayProps) {
+  useEffect(() => {
+    const previousOverflow = document.body.style.overflow;
+    document.body.style.overflow = "hidden";
+
+    const handleKeyDown = (event: KeyboardEvent) => {
+      if (event.key === "Escape") {
+        event.preventDefault();
+        onClose();
+      }
+    };
+
+    document.addEventListener("keydown", handleKeyDown);
+    return () => {
+      document.body.style.overflow = previousOverflow;
+      document.removeEventListener("keydown", handleKeyDown);
+    };
+  }, [onClose]);
+
+  return createPortal(
+    <div
+      role="dialog"
+      aria-modal="true"
+      aria-label={title}
+      className="fixed inset-0 z-[9999] flex flex-col bg-black/90 backdrop-blur-sm"
+      onClick={(event) => {
+        if (event.target === event.currentTarget) onClose();
+      }}
+    >
+      <div className="flex min-h-14 items-center justify-between gap-3 border-b border-glass-border bg-deep-black/85 px-3 py-2 sm:px-5">
+        <span className="min-w-0 truncate text-sm font-medium text-text-primary">{title}</span>
+        <div className="flex shrink-0 items-center gap-1.5">
+          <a
+            href={src}
+            target="_blank"
+            rel="noopener noreferrer"
+            className="inline-flex h-9 w-9 items-center justify-center rounded-lg text-text-secondary transition-colors hover:bg-white/10 hover:text-text-primary"
+            aria-label="Open original image"
+            onClick={(event) => event.stopPropagation()}
+          >
+            <ExternalLink className="h-4 w-4" />
+          </a>
+          <button
+            type="button"
+            onClick={(event) => {
+              event.stopPropagation();
+              onClose();
+            }}
+            className="inline-flex h-9 w-9 items-center justify-center rounded-lg text-text-secondary transition-colors hover:bg-white/10 hover:text-text-primary"
+            aria-label="Close image preview"
+          >
+            <X className="h-5 w-5" />
+          </button>
+        </div>
+      </div>
+      <div className="flex min-h-0 flex-1 items-center justify-center overflow-auto p-3 sm:p-6">
+        <img
+          src={src}
+          alt={title}
+          className="max-h-[calc(100vh-6.5rem)] max-w-full object-contain shadow-2xl"
+          onError={onImageError}
+        />
+      </div>
+    </div>,
+    document.body,
+  );
+}
